Fixed bug when reading and decoding options. Updated README. Code split.
The code was split tp dedicated resource files: device, graph, fifo

When decoding raw option data there was a bug which did not allocate te
correect amount of data in order for the options to be decoded properly.

The code has been split into dedicated source files for device, fifo and
graph
